refactor(shared): move mistral provider data to module
